refactor(odoo): extract action definitions
