+++ integrations/odoo/definitions/index.ts
@@ -1,0 +1,1 @@
+export { actions } from './actions'
